Basic encode module refactoring #121
